Split editor code in its own extension
- clean up command extensions
- add REST calls for the communication with the server
- register the generate workflow code in the navigator context menu
