+++ web/browser-app/package.json
@@ -22,9 +22,11 @@
     "@theia/outline-view": "latest",
     "coffee-editor-extension": "1.0.0",
     "coffee-workflow-analyzer": "0.0.0",
+    "coffee-workflow-analyzer-editors": "0.0.0",
+    "coffee-java-extension": "0.0.0",
     "glsp-theia-extension":"0.1.0",
-    "workflow-glsp-extension": "0.0.0",
-    "coffee-java-extension": "0.0.0"
+    "workflow-glsp-extension": "0.0.0"
+
   },
   "devDependencies": {
     "@theia/cli": "latest"
