--- web/coffee-java-extension/src/browser/language-contribution.ts
@@ -1,120 +1,0 @@
-import { injectable, inject } from 'inversify';
-import { Workspace, Languages, LanguageClientFactory, BaseLanguageClientContribution } from '@theia/languages/lib/browser';
-
-@injectable()
-export class WorkflowClientContribution extends BaseLanguageClientContribution {
-
-    readonly id = "wf";
-    readonly name = "WF";
-
-    constructor(
-        @inject(Workspace) protected readonly workspace: Workspace,
-        @inject(Languages) protected readonly languages: Languages,
-        @inject(LanguageClientFactory) protected readonly languageClientFactory: LanguageClientFactory
-    ) {
-        super(workspace, languages, languageClientFactory);
-    }
-
-    protected get globPatterns() {
-        return [
-            '**/*.wf'
-        ];
-    }
-}
-
-// register language with monaco on first load
-registerDSL();
-
-export function registerDSL() {
-    // initialize monaco
-    monaco.languages.register({
-        id: 'wf',
-        aliases: ['WF', 'wf'],
-        extensions: ['.wf'],
-        mimetypes: ['text/wf']
-    })
-    monaco.languages.setLanguageConfiguration('wf', {
-        comments: {
-            lineComment: "//",
-            blockComment: ['/*', '*/']
-        },
-        brackets: [['{', '}'], ['(', ')']],
-        autoClosingPairs: [
-            {
-                open: '{',
-                close: '}'
-            },
-            {
-                open: '(',
-                close: ')'
-            }]
-    })
-    monaco.languages.setMonarchTokensProvider('wf', <any>{
-        // Set defaultToken to invalid to see what you do not tokenize yet
-        // defaultToken: 'invalid',
-
-        keywords: [
-            'workflowModel' ,'probabilities', 'assertions','low' ,'medium','high'
-        ],
-
-        typeKeywords: [
-            'boolean', 'number', 'string',"int"
-        ],
-
-        operators: [
-            ':',"=",","
-        ],
-
-        // we include these common regular expressions
-        symbols: /[=><!~?:&|+\-*\/\^%]+/,
-        escapes: /\\(?:[abfnrtv\\"']|x[0-9A-Fa-f]{1,4}|u[0-9A-Fa-f]{4}|U[0-9A-Fa-f]{8})/,
-
-        // The main tokenizer for our languages
-        tokenizer: {
-            root: [
-                // identifiers and keywords
-                [/[a-z_$][\w$]*/, {
-                    cases: {
-                        '@typeKeywords': 'keyword',
-                        '@keywords': 'keyword',
-                        '@default': 'identifier'
-                    }
-                }],
-                [/[A-Z][\w\$]*/, 'type.identifier'],  // to show class names nicely
-
-                // whitespace
-                { include: '@whitespace' },
-
-                // delimiters and operators
-                [/[{}()\[\]]/, '@brackets'],
-                [/[<>](?!@symbols)/, '@brackets'],
-                [/@symbols/, {
-                    cases: {
-                        '@operators': 'operator',
-                        '@default': ''
-                    }
-                }]
-            ],
-
-            whitespace: [
-                [/[ \t\r\n]+/, 'white'],
-                [/\/\*/, 'comment', '@comment'],
-                [/\/\/.*$/, 'comment'],
-            ],
-
-            comment: [
-                [/[^\/*]+/, 'comment'],
-                [/\/\*/, 'comment.invalid'],
-                ["\\*/", 'comment', '@pop'],
-                [/[\/*]/, 'comment']
-            ],
-
-            string: [
-                [/[^\\"]+/, 'string'],
-                [/@escapes/, 'string.escape'],
-                [/\\./, 'string.escape.invalid'],
-                [/"/, 'string', '@pop']
-            ],
-        },
-    })
-}
