+++ web/coffee-workflow-analyzer-editors/src/browser/editor-contribution.ts
