+++ web/coffee-workflow-analyzer-editors/src/browser/monaco.d.ts
@@ -1,0 +1,1 @@
+/// <reference types='monaco-editor-core/monaco'/>
