+++ web/coffee-workflow-analyzer-editors/src/common/request-file-protocol.ts
