+++ web/coffee-workflow-analyzer/src/browser/command-contribution.ts
@@ -1,12 +1,12 @@
 import { CommandContribution, CommandRegistry, MenuContribution, MenuModelRegistry, Command, SelectionService } from "@theia/core/lib/common";
 import { injectable, inject } from "inversify";
-import { WorkflowClientContribution } from "./language-contribution";
-import { ExecuteCommandRequest, Workspace } from "@theia/languages/lib/common";
+
+import { Workspace } from "@theia/languages/lib/common";
 import { EditorManager, EDITOR_CONTEXT_MENU } from "@theia/editor/lib/browser";
-import { isString } from "util";
+
 import { UriAwareCommandHandler, UriCommandHandler } from "@theia/core/lib/common/uri-command-handler"
 import URI from "@theia/core/lib/common/uri";
-import { open, OpenerService } from "@theia/core/lib/browser";
+import { OpenerService } from "@theia/core/lib/browser";
 
 export const ANALYZE_COMMAND: Command = {
     id: "workflow.analyze.command",
@@ -17,8 +17,6 @@ export class WorkflowCommandContribution implements CommandContribution, MenuCon
 
 
     constructor(
-        @inject(WorkflowClientContribution)
-        protected readonly clientContributon: WorkflowClientContribution,
         @inject(Workspace)
         protected readonly workspace: Workspace,
         @inject(EditorManager)
@@ -36,22 +34,32 @@ export class WorkflowCommandContribution implements CommandContribution, MenuCon
     registerCommands(registry: CommandRegistry): void {
         registry.registerCommand(ANALYZE_COMMAND, this.newUriAwareCommandHandler({
             execute: async (uri) => {
-                const documentURL = uri.path.toString().replace("file://", "");
-                const client = await this.clientContributon.languageClient;
-                var result = await client.sendRequest(ExecuteCommandRequest.type, {
-                    command: "workflow.analyze",
-                    arguments: [documentURL]
-                });
-                if (isString(result)) {
-                    result = JSON.parse(result)
-                    if (result.error == undefined) {
-                        let fileName = result.name;
-                        let uri = new URI("file://" + fileName);
-                        open(this.openHandler, uri)
-                    }
+                //const documentURL = uri.path.toString().replace("file://", "");
 
+                // TODO communicate with the server via REST
+                const xhttp = new XMLHttpRequest();
+                xhttp.open("GET", "https://api.github.com/repos/vmg/redcarpet/issues?state=closed", true);
+                xhttp.send();
+
+                xhttp.onreadystatechange = (e) => {
+                    console.log(xhttp.responseText)
                 }
-                return result
+
+                // const client = await this.clientContributon.languageClient;
+                // var result = await client.sendRequest(ExecuteCommandRequest.type, {
+                //     command: "workflow.analyze",
+                //     arguments: [documentURL]
+                // });
+                // if (isString(result)) {
+                //     result = JSON.parse(result)
+                //     if (result.error == undefined) {
+                //         let fileName = result.name;
+                //         let uri = new URI("file://" + fileName);
+                //         open(this.openHandler, uri)
+                //     }
+
+                // }
+                // return result
             },
             isVisible: (uri) => uri.toString().endsWith("wfconfig"),
             isEnabled: (uri) => uri.toString().endsWith("wfconfig")
