+++ web/package.json
@@ -9,6 +9,6 @@
     "lerna": "2.4.0"
   },
   "workspaces": [
-    "coffee-editor-extension", "coffee-java-extension", "coffee-workflow-analyzer", "workflow-glsp-extension", "workflow-sprotty", "browser-app", "electron-app"
+    "coffee-editor-extension", "coffee-java-extension", "coffee-workflow-analyzer", "coffee-workflow-analyzer-editors", "workflow-glsp-extension", "workflow-sprotty", "browser-app", "electron-app"
   ]
 }
